Update: Modularise Empty component
