+++ src/components/alexandria/Empty/styles.scss
@@ -1,6 +1,6 @@
-@import '../color';
-@import '../font-size';
-@import '../variable';
+@import 'styles/color';
+@import 'styles/font-size';
+@import 'styles/variable';
 
 .empty-component {
   cursor: default;
